requirements.txt
streamlit>=1.37.0
pandas>=2.2.0
tweepy>=4.14.0
nltk>=3.9
plotly>=5.22.0
wordcloud>=1.9.3
python-dotenv>=1.0.1
requests>=2.32.0
